+++ src/aux-records/NotificationMessenger.ts
@@ -2,6 +2,8 @@ import { StudioComIdRequest } from './RecordsStore';
 import { UserInstReport } from './ModerationStore';
 import { DateTime } from 'luxon';
 import { z } from 'zod';
+import { ModerationFileScanLabel } from './ModerationJobProvider';
+import { ResourceKinds } from '@casual-simulation/aux-common';
 
 /**
  * Defines an interface for a class that is able to send records notifications.
@@ -58,6 +60,7 @@ export class MultiNotificationMessenger implements NotificationMessenger {
 
 export type RecordsNotification =
     | UserInstReportNotification
+    | ModerationResourceScanNotification
     | StudioComIdRequestNotification;
 
 /**
@@ -100,6 +103,52 @@ export interface UserInstReportNotification extends ResourceNotification {
     report: UserInstReport;
 }
 
+/**
+ * Defines a notification that is for a user inst report.
+ */
+export interface ModerationResourceScanNotification
+    extends ResourceNotification {
+    /**
+     * The kind of the resource that was scanned.
+     */
+    resource: ResourceKinds;
+
+    /**
+     * The name of the record that the resource was scanned in.
+     */
+    recordName: string | null;
+
+    /**
+     * The ID of the resource that was scanned.
+     */
+    resourceId: string;
+
+    /**
+     * The ID of the moderation result that was created for the resource.
+     */
+    resultId: string;
+
+    /**
+     * The labels that were detected in the file.
+     */
+    labels: ModerationFileScanLabel[];
+
+    /**
+     * The time of the scan in unix time in milliseconds.
+     */
+    timeMs: number;
+
+    /**
+     * The message that should be sent with the notification.
+     */
+    message: string;
+
+    /**
+     * The label that caused the resource to be banned.
+     */
+    bannedLabel?: ModerationFileScanLabel;
+}
+
 export interface StudioComIdRequestNotification extends ResourceNotification {
     resource: 'studio_com_id_request';
 
@@ -109,7 +158,11 @@ export interface StudioComIdRequestNotification extends ResourceNotification {
     request: StudioComIdRequest;
 }
 
-export type NotificationResourceActions = 'created' | 'updated' | 'deleted';
+export type NotificationResourceActions =
+    | 'created'
+    | 'updated'
+    | 'deleted'
+    | 'scanned';
 
 export const slackSchema = z.object({
     webhookUrl: z
@@ -181,6 +234,13 @@ export function formatNotificationAsString(
             return formatUserInstReportNotificationAsString(notification);
         case 'studio_com_id_request':
             return formatStudioComIdRequestNotificationAsString(notification);
+        case 'file':
+        case 'data':
+        case 'event':
+        case 'inst':
+            return formatResourceNotificationAsString(notification);
+        default:
+            return JSON.stringify(notification, undefined, 2);
     }
 }
 
@@ -222,4 +282,34 @@ Requested comID: ${notification.request.requestedComId}
 Time: ${time}
 Reporting User: ${notification.request.userId ?? '(null)'}
 Requesting IP: ${notification.request.requestingIpAddress ?? '(null)'}`;
+}
+
+export function formatResourceNotificationAsString(
+    notification: ModerationResourceScanNotification
+): string {
+    const time = DateTime.fromMillis(notification.timeMs, {
+        zone: 'utc',
+    }).toISO();
+    return `A ${notification.resource} was ${
+        notification.action
+    } for moderation labels.
+
+Message: ${notification.message}
+RecordName: ${notification.recordName}
+FileName: ${notification.resourceId}
+ResultId: ${notification.resultId}
+BannedLabel: ${
+        notification.bannedLabel
+            ? formatLabel(notification.bannedLabel)
+            : '(null)'
+    }
+Labels: 
+${notification.labels.map(formatLabel).join('\n')}
+Time: ${time}`;
+
+    function formatLabel(label: ModerationFileScanLabel): string {
+        return `- ${label.name}${label.category ? ':' + label.category : ''} (${
+            label.confidence
+        })`;
+    }
 }