+++ src/aux-runtime/runtime/Transpiler.ts
@@ -82,7 +82,10 @@ export interface TranspilerMacro {
     replacement: (val: string) => string;
 }
 
-const TypeScriptVisistorKeys: { [nodeType: string]: string[] } = {
+/**
+ * The estraverse visitor keys that are used for TypeScript nodes.
+ */
+export const TypeScriptVisistorKeys: { [nodeType: string]: string[] } = {
     TSTypeParameterDeclaration: [],
     TSCallSignatureDeclaration: [],
     TSConstructSignatureDeclaration: [],
@@ -91,6 +94,7 @@ const TypeScriptVisistorKeys: { [nodeType: string]: string[] } = {
     TSEnumDeclaration: [],
     TSTypeAliasDeclaration: [],
     TSDeclareFunction: [],
+    TSDeclareMethod: [],
 
     TSExternalModuleReference: [],
     TSQualifiedName: [],
@@ -142,6 +146,7 @@ const TypeScriptVisistorKeys: { [nodeType: string]: string[] } = {
     TSMethodSignature: [],
     TSPropertySignature: [],
     TSAsExpression: [],
+    TSParameterProperty: ['parameter'],
 
     ClassDeclaration: [
         ...VisitorKeys.ClassDeclaration,
@@ -495,6 +500,18 @@ export class Transpiler {
 
                     if (n.abstract === true) {
                         this._removeClassAbstract(n, doc, text);
+                    }
+                } else if (n.type === 'TSParameterProperty') {
+                    // do nothing
+                } else if (n.type === 'MethodDefinition') {
+                    if (n.abstract) {
+                        this._removeNodeOrReplaceWithUndefined(n, doc, text);
+                    } else if (n.accessibility) {
+                        this._removeAccessibility(n, doc, text);
+                    }
+                } else if (n.type === 'PropertyDefinition') {
+                    if (n.accessibility) {
+                        this._removeAccessibility(n, doc, text);
                     }
                 } else if (n.type === 'TSAsExpression') {
                     this._removeAsExpression(n, doc, text);
@@ -1227,6 +1244,37 @@ export class Transpiler {
             abstractStart.index,
             abstractEnd.index - abstractStart.index
         );
+    }
+
+    private _removeAccessibility(node: any, doc: Doc, text: Text): any {
+        doc.clientID += 1;
+        const version = { '0': getClock(doc, 0) };
+
+        const accessibility: string = node.accessibility + ' ';
+        const t = text.toString();
+
+        const relativeStart = createRelativePositionFromStateVector(
+            text,
+            version,
+            node.start,
+            -1,
+            true
+        );
+        const absoluteStart = createAbsolutePositionFromRelativePosition(
+            relativeStart,
+            doc
+        );
+
+        const indexOfAccessibility = t.indexOf(
+            accessibility,
+            absoluteStart.index
+        );
+
+        if (indexOfAccessibility < 0 || indexOfAccessibility > node.key.start) {
+            return;
+        }
+
+        text.delete(indexOfAccessibility, accessibility.length);
     }
 
     private _removeAsExpression(node: any, doc: Doc, text: Text): any {