+++ src/casualos-cli/config.ts
@@ -1,0 +1,4 @@
+export interface CliConfig {
+    get(key: string): any;
+    set(key: string, value: any): void;
+}