+++ src/casualos-cli/infra-utils.ts
@@ -1,0 +1,3 @@
+export function getRepoName(path: string) {
+    return path.replace(/^(?:\w:|~)?[\.\/\\]*/g, '').replace(/[\/\\]/g, '-');
+}